财务

+++ hotPot/src/com/hot/model/Finance.java
@@ -2,10 +2,10 @@ package com.hot.model;
 
 public class Finance {
 		public int fid;
-		public int fincome;
-		public int fexpend;
-		public int fprofit;
-		public int fbalance;
+		public double fincome;
+		public double fexpend;
+		public double fprofit;
+		public double fbalance;
 		public String ftime;
 		public int getFid() {
 			return fid;
@@ -13,29 +13,29 @@ public class Finance {
 		public void setFid(int fid) {
 			this.fid = fid;
 		}
-		public int getFincome() {
+		public double getFincome() {
 			return fincome;
 		}
-		public void setFincome(int fincome) {
+		public void setFincome(double fincome) {
 			this.fincome = fincome;
 		}
-		public int getFexpend() {
+		public double getFexpend() {
 			return fexpend;
 		}
-		public void setFexpend(int fexpend) {
+		public void setFexpend(double fexpend) {
 			this.fexpend = fexpend;
 		}
-		public int getFprofit() {
+		public double getFprofit() {
 			return fprofit;
 		}
-		public void setFprofit(int fprofit) {
+		public void setFprofit(double fprofit) {
 			this.fprofit = fprofit;
 		}
 
-		public int getFbalance() {
+		public double getFbalance() {
 			return fbalance;
 		}
-		public void setFbalance(int fbalance) {
+		public void setFbalance(double fbalance) {
 			this.fbalance = fbalance;
 		}
 		public String getFtime() {
